server/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sse-starlette>=2.1.0
websocket-client>=1.8.0
requests>=2.32.0
python-dotenv>=1.0.0
pydantic>=2.9.0
